[Jishnu/Anjaly] - Added function to both team and player prototype
What changes have you made?
Added a function in team's prototype to check a team has a pair or not
*

Why do we need this change?
During game we need to know a team has pair inorder to change bidpoint
*

What can this change impact?
pair will increase/decrease the bidpoint acording to the game
*